backend/requirements.txt
Django>=4.2
djangorestframework>=3.15
celery>=5.4
redis>=5.0
dj-database-url>=2.2
psycopg[binary]>=3.2
python-dotenv>=1.0
sqlalchemy>=2.0
PyJWT>=2.9
django-cors-headers>=4.4
